syntax: consolidate function parsing in `item.rs`
